Setup bundler for use with Tracks
Add the Gemfile and Gemfile.lock files used to specify the dependencies
of the app.

The Gemfile specifies the dependencies and the Gemfile.lock is a
snapshot of the dependencies. If the Gemfile changes, then Bundler will
look at the differences between the Gemfile and the Gemfile.lock to
determine what other gems to install or upgrade