Update the README for developers.
We're using bundler now so update the piece about installing the
dependencies.

Also do some reformatting at the 80 column boundary in order to avoid
having really long lines that don't look quite so nice in a smaller
editor window.

Add information about running the Test::Unit tests.

Remove the reference to the forum since it's not accepting new
registrations anymore and users are advised to use the mailing list
elsewhere.