Swap min/max variables in i18n translation
The :min was pointing at the max value, and the inverse.

I had to tweak a few of the language files as well, to make sure that
minimum and maximum are pointing to the correct values.

I didn't touch the hebrew language file, so someone with the correct
language skills may want to take a look at it to make sure that the text
matches the interpolation variables that get sent in.