Completed todos are now selectable.
So far, the whole row with a todo was draggable, with handle grip_span.
This means for normal todos the whole action is draggable, but user
must drag it by clicking on grip_span.

Grip span has not been rendered for completed todos, so the whole row
could be dragged, preventing user from selecting any text in completed
todo.

This commit makes grip_span rendered also for completed tasks, allowing
them to be dragged by grip_span, but the rest of the text is selectable
like for normal todos.